Bump the version to 0.1.0
Having the a version of 0.0.x doesn't allow users to get patch releases
very easily. In SemVer if you specify a caret before the version then it
will allow only updates to the subsequent numbers, i.e. ^x.y.z will allow
updates of y and z, ^0.x.y allows updates of y, but ^0.0.x does not allow
any updates.

This commit changes the version to 0.1.0. This is probably an appropriate
starting point as the module is now fully functional but could do with a
patch for including modules that are retrieved from the root directory,
there is a pending test for this.