Fix test path expectations
The literal test path expectations were only valid when running on a *nix system and
not when running on Windows.

I have added a function to normalize the forward slashes in the path literal to Node's
path separator (path.sep) to ensure that the expected paths are now correct for the OS
where the tests are run.